+++ menu/cbs/menu_cbs_ok.c
@@ -1,6 +1,7 @@
 /*  RetroArch - A frontend for libretro.
  *  Copyright (C) 2011-2017 - Daniel De Matteis
  *  Copyright (C) 2016-2017 - Brad Parker
+ *  Copyright (C) 2015-2017 - Andrés Suárez
  *
  *  RetroArch is free software: you can redistribute it and/or modify it under the terms
  *  of the GNU General Public License as published by the Free Software Found-
@@ -2090,10 +2091,12 @@ static int action_ok_cheat_file_save_as(const char *path,
 enum
 {
    ACTION_OK_REMAP_FILE_SAVE_CORE = 0,
-   ACTION_OK_REMAP_FILE_SAVE_GAME
+   ACTION_OK_REMAP_FILE_SAVE_GAME,
+   ACTION_OK_REMAP_FILE_REMOVE_CORE,
+   ACTION_OK_REMAP_FILE_REMOVE_GAME
 };
 
-static int generic_action_ok_remap_file_save(const char *path,
+static int generic_action_ok_remap_file_operation(const char *path,
       const char *label, unsigned type, size_t idx, size_t entry_idx,
       unsigned action_type)
 {
@@ -2118,10 +2121,12 @@ static int generic_action_ok_remap_file_save(const char *path,
    switch (action_type)
    {
       case ACTION_OK_REMAP_FILE_SAVE_CORE:
+      case ACTION_OK_REMAP_FILE_REMOVE_CORE:
          if (!string_is_empty(core_name))
             fill_pathname_join(file, core_name, core_name, sizeof(file));
          break;
       case ACTION_OK_REMAP_FILE_SAVE_GAME:
+      case ACTION_OK_REMAP_FILE_REMOVE_GAME:
          if (!string_is_empty(core_name))
             fill_pathname_join(file, core_name,
                   path_basename(path_get(RARCH_PATH_BASENAME)), sizeof(file));
@@ -2131,30 +2136,72 @@ static int generic_action_ok_remap_file_save(const char *path,
    if(!path_file_exists(directory))
        path_mkdir(directory);
 
-   if(input_remapping_save_file(file))
-      runloop_msg_queue_push(
-            msg_hash_to_str(MSG_REMAP_FILE_SAVED_SUCCESSFULLY),
-            1, 100, true);
-   else
-      runloop_msg_queue_push(
-            msg_hash_to_str(MSG_ERROR_SAVING_REMAP_FILE),
-            1, 100, true);
+   if (action_type < ACTION_OK_REMAP_FILE_REMOVE_CORE)
+   {
+      if(input_remapping_save_file(file))
+      {
+         if (action_type == ACTION_OK_REMAP_FILE_SAVE_CORE)
+            rarch_ctl(RARCH_CTL_SET_REMAPS_CORE_ACTIVE, NULL);
+         else
+            rarch_ctl(RARCH_CTL_SET_REMAPS_GAME_ACTIVE, NULL);
 
+         runloop_msg_queue_push(
+               msg_hash_to_str(MSG_REMAP_FILE_SAVED_SUCCESSFULLY),
+               1, 100, true);
+      }
+      else
+         runloop_msg_queue_push(
+               msg_hash_to_str(MSG_ERROR_SAVING_REMAP_FILE),
+               1, 100, true);
+   }
+   else
+   {
+      RARCH_LOG("removing %s", file);
+      if(input_remapping_remove_file(file))
+      {
+         if (action_type == ACTION_OK_REMAP_FILE_REMOVE_CORE)
+            rarch_ctl(RARCH_CTL_UNSET_REMAPS_CORE_ACTIVE, NULL);
+         else
+            rarch_ctl(RARCH_CTL_UNSET_REMAPS_GAME_ACTIVE, NULL);
+
+         runloop_msg_queue_push(
+               msg_hash_to_str(MSG_REMAP_FILE_REMOVED_SUCCESSFULLY),
+               1, 100, true);
+      }
+      else
+         runloop_msg_queue_push(
+               msg_hash_to_str(MSG_ERROR_REMOVING_REMAP_FILE),
+               1, 100, true);
+   }
    return 0;
 }
 
 static int action_ok_remap_file_save_core(const char *path,
       const char *label, unsigned type, size_t idx, size_t entry_idx)
 {
-   return generic_action_ok_remap_file_save(path, label, type,
+   return generic_action_ok_remap_file_operation(path, label, type,
          idx, entry_idx, ACTION_OK_REMAP_FILE_SAVE_CORE);
 }
 
 static int action_ok_remap_file_save_game(const char *path,
       const char *label, unsigned type, size_t idx, size_t entry_idx)
 {
-   return generic_action_ok_remap_file_save(path, label, type,
+   return generic_action_ok_remap_file_operation(path, label, type,
          idx, entry_idx, ACTION_OK_REMAP_FILE_SAVE_GAME);
+}
+
+static int action_ok_remap_file_remove_core(const char *path,
+      const char *label, unsigned type, size_t idx, size_t entry_idx)
+{
+   return generic_action_ok_remap_file_operation(path, label, type,
+         idx, entry_idx, ACTION_OK_REMAP_FILE_REMOVE_CORE);
+}
+
+static int action_ok_remap_file_remove_game(const char *path,
+      const char *label, unsigned type, size_t idx, size_t entry_idx)
+{
+   return generic_action_ok_remap_file_operation(path, label, type,
+         idx, entry_idx, ACTION_OK_REMAP_FILE_REMOVE_GAME);
 }
 
 int action_ok_path_use_directory(const char *path,
@@ -4765,6 +4812,12 @@ static int menu_cbs_init_bind_ok_compare_label(menu_file_list_cbs_t *cbs,
             break;
          case MENU_ENUM_LABEL_REMAP_FILE_SAVE_GAME:
             BIND_ACTION_OK(cbs, action_ok_remap_file_save_game);
+            break;
+         case MENU_ENUM_LABEL_REMAP_FILE_REMOVE_CORE:
+            BIND_ACTION_OK(cbs, action_ok_remap_file_remove_core);
+            break;
+         case MENU_ENUM_LABEL_REMAP_FILE_REMOVE_GAME:
+            BIND_ACTION_OK(cbs, action_ok_remap_file_remove_game);
             break;
          case MENU_ENUM_LABEL_CONTENT_COLLECTION_LIST:
             BIND_ACTION_OK(cbs, action_ok_content_collection_list);