+++ CHANGES.md
@@ -7,6 +7,7 @@
 - INPUT: allow specifying analog dpad mode in remap files
 - INPUT: allow saving libretro device to remap files
 - INPUT: allow saving analog dpad mode to remap files
+- INPUT: allow removing core and game remap files from the menu
 - COMMON: Add 'Delete Core'option to Core Information menu.
 - COMMON: Allow Max Timing Skew to be set to 0.
 - LOCALIZATION: Update Russian translation